TIKA-4271 Support tcpSocket probe
With the active gRPC Server development
(https://issues.apache.org/jira/browse/TIKA-4181), there is need to be
able to deploy this Helm chart against the running gRPC service.

As the exposed gRPC service cannot use the HTTP probe, we must adapt the
Helm chart to be configurable for a tcp socket

This change identifies `tcp` in the scheme variable of either probe and
ensures the appropriate mechanism is used.

Existing HTTP(S) functionality continues to be supported.